+++ src/components/PaymentInfo/PaymentInfo.stories.tsx
@@ -1,10 +1,10 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { reactRouterParameters, withRouter } from 'storybook-addon-react-router-v6';
-import OrderInfo from './OrderInfo';
+import PaymentInfo from './PaymentInfo.tsx';
 
 const meta = {
-  title: 'OrderInfo',
-  component: OrderInfo,
+  title: 'PaymentInfo',
+  component: PaymentInfo,
   decorators: [withRouter],
   parameters: {
     reactRouter: reactRouterParameters({
@@ -18,7 +18,7 @@ const meta = {
       routing: { path: '/order-info' },
     }),
   },
-} satisfies Meta<typeof OrderInfo>;
+} satisfies Meta<typeof PaymentInfo>;
 
 export default meta;
