+++ src/components/PaymentInfo/styled.tsx
